Salvages the current server as a starting point.
