+++ packages/reactotron-cli/.gitignore
@@ -1,0 +1,5 @@
+node_modules
+npm-debug.log
+coverage
+.nyc_output
+dist
