+++ packages/reactotron-cli/package.json
@@ -1,0 +1,57 @@
+{
+  "name": "reactotron-cli",
+  "version": "0.0.1",
+  "description": "Reactotron terminal mode engage!",
+  "bin": {
+    "reactotron": "./dist/reactotron.js"
+  },
+  "scripts": {
+    "test": "ava",
+    "watch": "ava --watch",
+    "coverage": "nyc ava",
+    "start": "babel-node src/index.js",
+    "build": "rollup -c rollup.config.js"
+  },
+  "repository": "https://github.com/skellock/reactotron/tree/master/packages/reactotron-cli",
+  "author": "Steve Kellock",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/skellock/reactotron/issues"
+  },
+  "homepage": "https://reactotron.com",
+  "files": [
+    "dist",
+    "LICENSE",
+    "README.md"
+  ],
+  "devDependencies": {
+    "ava": "^0.15.2",
+    "babel-core": "^6.11.4",
+    "babel-eslint": "^6.1.2",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-es2015-rollup": "^1.1.1",
+    "babel-preset-stage-1": "^6.5.0",
+    "nyc": "^7.1.0",
+    "rollup": "^0.34.1",
+    "rollup-plugin-babel": "^2.6.1",
+    "standard": "^7.1.2",
+    "blessed": "^0.1.81",
+    "gemoji": "^1.0.0",
+    "moment": "^2.12.0",
+    "strman": "^1.0.1",
+    "reactotron-core-server": "^0.0.1"
+  },
+  "ava": {
+    "require": [
+      "babel-core/register"
+    ],
+    "babel": "inherit"
+  },
+  "standard": {
+    "parser": "babel-eslint"
+  },
+  "dependencies": {
+    "ramda": "^0.21.0",
+    "ramdasauce": "^1.1.0"
+  }
+}
