+++ packages/reactotron-cli/rollup.config.js
@@ -1,0 +1,13 @@
+import babel from 'rollup-plugin-babel'
+
+export default {
+  entry: 'src/index.js',
+  format: 'cjs',
+  plugins: [
+    babel({
+      babelrc: false,
+      presets: ['es2015-rollup', 'stage-1']
+    })
+  ],
+  dest: 'dist/index.js'
+}
