+++ packages/reactotron-cli/src/commands/client/clientAdd.js
