+++ packages/reactotron-cli/src/commands/client/clientRemove.js
