+++ packages/reactotron-cli/src/commands/console/consoleError.js
