+++ packages/reactotron-cli/src/commands/content/contentClear.js
