+++ packages/reactotron-cli/src/commands/menus/menuMain.js
