+++ packages/reactotron-cli/src/commands/program/die.js
