+++ packages/reactotron-cli/src/commands/redux/reduxSubscribeDelete.js
