+++ packages/reactotron-cli/src/commands/redux/reduxSubscribeDeletePrompt.js
