+++ packages/reactotron-cli/src/commands/redux/reduxValueRequest.js
