+++ packages/reactotron-cli/src/commands/redux/reduxValueResponse.js
